requirements.txt
Pillow==10.0.1
PySimpleGui==4.60.5